Created own ideadetails component
