+++ src/components/SpecialistPortal/Dashboard/AssessmentSection/AssessmentSection.css
@@ -29,20 +29,8 @@
 }
 
 .assessments-content {
+  --origin: top;
   animation-name: slide;
   animation-timing-function: ease-in-out;
   animation-duration: 0.3s;
-}
-
-@keyframes slide {
-  0% {
-    transform: scaleY(0);
-    transform-origin: top;
-    opacity: 0;
-  }
-  100% {
-    transform: scaleY(1);
-    transform-origin: top;
-    opacity: 1;
-  }
 }
