+++ src/components/SpecialistPortal/Dashboard/CommentSection/CommentSection.css
@@ -1,18 +1,6 @@
 .commentsectioncontent {
+  --origin: top;
   animation-name: slide;
   animation-timing-function: ease-in-out;
   animation-duration: 0.3s;
-}
-
-@keyframes slide {
-  0% {
-    transform: scaleY(0);
-    transform-origin: top;
-    opacity: 0;
-  }
-  100% {
-    transform: scaleY(1);
-    transform-origin: top;
-    opacity: 1;
-  }
 }
