+++ src/components/reogranisation/IdeaDetails/IdeaDetails.css
@@ -1,0 +1,30 @@
+.ideadetailscontent {
+  --origin: top;
+  animation-name: slide;
+  animation-timing-function: ease-in-out;
+  animation-duration: 0.3s;
+}
+
+.ideadetailscontentsummary {
+  --origin: bottom;
+  animation-name: slide;
+  animation-timing-function: ease-in-out;
+  animation-duration: 0.3s;
+}
+
+@keyframes slide {
+  0% {
+    transform: scaleY(0);
+    transform-origin: var(--origin);
+    opacity: 0;
+  }
+  100% {
+    transform: scaleY(1);
+    transform-origin: var(--origin);
+    opacity: 1;
+  }
+}
+
+.ideadetailscontentsummary ul {
+  list-style: none;
+}
